+++ server/Mac/mf_input.h
@@ -22,7 +22,7 @@
 
 #include "mf_interface.h"
 
-BOOL mf_input_keyboard_event(rdpInput* input, UINT16 flags, UINT16 code);
+BOOL mf_input_keyboard_event(rdpInput* input, UINT16 flags, UINT8 code);
 BOOL mf_input_unicode_keyboard_event(rdpInput* input, UINT16 flags, UINT16 code);
 BOOL mf_input_mouse_event(rdpInput* input, UINT16 flags, UINT16 x, UINT16 y);
 BOOL mf_input_extended_mouse_event(rdpInput* input, UINT16 flags, UINT16 x, UINT16 y);
